simplified the extension loading mechanism, fixed a bug in the main executable.
It's now up to the compile-time extension code to return a 'require'
instruction AST if it needs a runtime part. By convention, runtime
parts are in the lib/extension directory, with the same name as their
compile-time counterparts suffixed with "-runtime".

In the metalua executable, the files where compiled into functions to
isolate their respective fenvs. This prevented them from returning
values, which was an issue for requirable modules. Fixed: the last
file's AST is inserted into a `Return{ }. Maybe this should be done to
the last chunk, irrespective of its origin?