spelling fixes to the documentation
Metalua is a substantial enhancement to Lua.  It deserves documentation
that is easily digestable.  People like me tend to be irritated by
typos, which might slow down casual readers.

No changes to any code, only the documentation is considered.  The
changes are in a separate branch "spott-fixspelling" tracking
"origin/master".